@@ -5,7 +5,7 @@ homepage: https://github.com/giann/bullseye
 publish_to: none
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   darty_json: ^1.0.0
